add jwt to the project

+++ controller/api/v1/posts_api.js
@@ -24,22 +24,23 @@ module.exports.destroy = async (req,res)=>{
     try{
         let post = await Post.findById(req.params.id);
     
-       // if(post.user == req.user.id){
+       if(post.user == req.user.id){
             post.remove();
 
-        await Comment.deleteMany({post : req.params.id});
+            await Comment.deleteMany({post : req.params.id});
        
-       return res.json(200,{
-           message:"Post and associated comments deleted successfully !"
-       });
-    }
-        // else{
-        //     req.flash('success','You cannot delete that post !');
-        //     return res.redirect('back');
-        // }
+            return res.json(200,{
+                message:"Post and associated comments deleted successfully !"
+            });
+        }
+        else{
+            return res.json(401,{
+                message:"You cannot delete this post !"
+            });
+        }
 
 
-    catch(err){
+    }catch(err){
         console.log("*****",err);
         return res.json(500,{
             message:"Internal Server Error !"
